+++ java/src/main/java/io/cucumber/java/JavaDefaultDataTableEntryTransformerDefinition.java
@@ -9,9 +9,7 @@ import io.cucumber.datatable.TableEntryByTypeTransformer;
 import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.regex.Pattern;
 
 import static io.cucumber.java.InvalidMethodSignatureException.builder;
 
@@ -106,46 +104,4 @@ class JavaDefaultDataTableEntryTransformerDefinition extends AbstractGlueDefinit
         return Invoker.invoke(lookup.getInstance(method.getDeclaringClass()), method, args);
     }
 
-    static class CamelCaseStringConverter {
-        private static final String WHITESPACE = " ";
-        private static final Pattern WHITESPACE_PATTERN = Pattern.compile("\\s+");
-
-        Map<String, String> toCamelCase(Map<String, String> fromValue) {
-            Map<String, String> newMap = new HashMap<>();
-            CamelCaseStringConverter converter = new CamelCaseStringConverter();
-            for (Map.Entry<String, String> entry : fromValue.entrySet()) {
-                newMap.put(converter.toCamelCase(entry.getKey()), entry.getValue());
-            }
-            return newMap;
-        }
-
-        String toCamelCase(String string) {
-            String[] parts = normalizeSpace(string).split(WHITESPACE);
-            parts[0] = uncapitalize(parts[0]);
-            for (int i = 1; i < parts.length; i++) {
-                parts[i] = capitalize(parts[i]);
-            }
-            return join(parts);
-        }
-
-        private String join(String[] parts) {
-            StringBuilder sb = new StringBuilder();
-            for (String s : parts) {
-                sb.append(s);
-            }
-            return sb.toString();
-        }
-
-        private String normalizeSpace(String originalHeaderName) {
-            return WHITESPACE_PATTERN.matcher(originalHeaderName.trim()).replaceAll(WHITESPACE);
-        }
-
-        private String capitalize(String string) {
-            return Character.toTitleCase(string.charAt(0)) + string.substring(1);
-        }
-
-        private String uncapitalize(String string) {
-            return Character.toLowerCase(string.charAt(0)) + string.substring(1);
-        }
-    }
 }
