+++ java/src/test/java/io/cucumber/java/JavaDefaultDataTableEntryTransformerDefinitionTest.java
@@ -2,7 +2,6 @@ package io.cucumber.java;
 
 import io.cucumber.core.backend.Lookup;
 import io.cucumber.datatable.TableCellByTypeTransformer;
-import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Method;
@@ -10,20 +9,15 @@ import java.lang.reflect.Type;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static java.util.Collections.singletonMap;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.core.Is.is;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class JavaDefaultDataTableEntryTransformerDefinitionTest {
 
     private final Map<String, String> fromValue = singletonMap("key", "value");
-    private final JavaDefaultDataTableEntryTransformerDefinition.CamelCaseStringConverter camelCaseConverter
-        = new JavaDefaultDataTableEntryTransformerDefinition.CamelCaseStringConverter();
     private final Lookup lookup = new Lookup() {
 
         @Override
@@ -135,26 +129,6 @@ class JavaDefaultDataTableEntryTransformerDefinitionTest {
     void method_must_have_cell_transformer_as_optional_third_argument() throws Throwable {
         Method method = JavaDefaultDataTableEntryTransformerDefinitionTest.class.getMethod("invalid_optional_third_type", Map.class, Type.class, String.class);
         assertThrows(InvalidMethodSignatureException.class, () -> new JavaDefaultDataTableEntryTransformerDefinition(method, lookup));
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"testString", "TestString", "Test String", "test String", "Test string"})
-    void convert_to_camel_case(String header) throws Throwable {
-        String expectedHeader = "testString";
-        String actualHeader = camelCaseConverter.toCamelCase(header);
-        Assert.assertThat(actualHeader, equalTo(expectedHeader));
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {
-        "threeWordsString", "ThreeWordsString", "three Words String", "Three Words String", "Three words String",
-        "Three Words string", "Three words string", "three Words string", "three words String", "threeWords string",
-        "three WordsString", "three wordsString",
-    })
-    void convert_three_words_to_camel_case(String header) throws Throwable {
-        String expectedHeader = "threeWordsString";
-        String actualHeader = camelCaseConverter.toCamelCase(header);
-        Assert.assertThat(actualHeader, equalTo(expectedHeader));
     }
 
     public Object invalid_optional_third_type(Map<String, String> fromValue, Type toValueType, String cellTransformer) {
